Fix: Cypress 10 now runs our channel create script to the point of login
There were package conflicts with eslint for cypress 10 and mocha,
so I removed all of the eslint from the repo. That's probably a bad
idea long term.

I've now copied enough of the nodejs_sdk tests to clean out the
account before we start a test. I've also found out that
initializing the client in the cypress commands _ALSO_ loggs you
into the webapp because cypress stuff all executes in the browser.